+++ src/ska/contrib/django/ska/exceptions.py
@@ -2,9 +2,5 @@ __title__ = 'ska.contrib.django.ska.exceptions'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2013 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
-__all__ = ('ImproperlyConfigured',)
+#__all__ = ('',)
 
-class ImproperlyConfigured(Exception):
-    """
-    Exception raised when developer didn't configure the code properly.
-    """